Display labels for fields with initial values.
Previously, I simply hid all the labels to start. That's fine for a blank form
but for forms that come pre-populated that doesn't work well.

Unrelated, I added another call to `_getCorrespondingField`. I feel like I do
that a lot and it seems inefficient to keep doing that. It might be good to
introduce a data structure instead of looking it up several times.

The alternative is to combine all the relevant code into one place but that
make the code less modular.